verity: fix calculations of partitions bigger than 4GiB
BUG=b:278966999
TEST=build and flash test image on draco

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4472233
Tested-by: YH Lin <yueherngl@chromium.org>
Auto-Submit: Dominik Behr <dbehr@chromium.org>
Owners-Override: Lalith Rajendran <lalithkraj@google.com>
Commit-Queue: YH Lin <yueherngl@chromium.org>

+++ verity/dm-bht.cc
@@ -268,7 +268,7 @@ int dm_bht_initialize_entries(struct dm_bht* bht) {
    * independently from the bht data structures.  Logically, the root is
    * depth=-1 and the block layer level is depth=bht->depth
    */
-  unsigned int last_index =
+  uint64_t last_index =
       base::bits::AlignUp(bht->block_count, bht->node_count) - 1;
   struct dm_bht_level* level = NULL;
   int depth;
@@ -277,8 +277,8 @@ int dm_bht_initialize_entries(struct dm_bht* bht) {
    * on allocation or sector calculation.
    */
   if (((last_index >> bht->node_count_shift) + 1) >
-      UINT_MAX / std::max((unsigned int)sizeof(struct dm_bht_entry),
-                          (unsigned int)to_sector(PAGE_SIZE))) {
+      UINT_MAX / std::max((uint64_t)sizeof(struct dm_bht_entry),
+                          (uint64_t)to_sector(PAGE_SIZE))) {
     LOG(ERROR) << "required entries " << last_index + 1 << " is too large.";
     return -EINVAL;
   }
